Rename models for more clarity and consistency
