+++ DNAnet/models/classification/peak_classification.py
@@ -9,12 +9,12 @@ from torchmetrics import Metric
 from DNAnet.data.data_models.extracted_peak import ExtractedPeak
 from DNAnet.data.data_models.hid_image import HIDImage
 from DNAnet.data.preprocessing.peak_utils import MARKER_TO_IDX
-from DNAnet.models.HIDImageBaseModel import HIDImageBaseModel
+from DNAnet.models.base_model import BaseModel
 from DNAnet.models.classification.peak_classification_torch import PeakClassificationModel
 from DNAnet.models.prediction import Prediction
 
 
-class PeakClassification(HIDImageBaseModel):
+class PeakClassification(BaseModel):
     def __init__(self,
                  labels: List[str],
                  device: Optional[str] = None,
